Split validation results into its own dock widget
Allowing it to be displayed at the same time as the redistrict
stats window

Fixes #70
